feat: rename /option -> /builder
